+++ src/engines/L3/element.js
@@ -240,6 +240,10 @@ const Props = {
   set textAlign(v) {
     this._props.textAlign = v
   },
+  set rtt(v) {
+    this._props.rtt = v
+    this._set.add('rtt')
+  },
 }
 
 const Element = {
@@ -271,9 +275,12 @@ const Element = {
     // correct for default white nodes (but not for text nodes)
     if (!props.__textnode && !this.props._set.has('color')) {
       this.props._props.color =
-        this.props._set.has('src') || this.props._set.has('texture') ? 0xffffffff : 0
+        this.props._set.has('src') ||
+        this.props._set.has('texture') ||
+        (this.props._set.has('rtt') && this.props._props.rtt === true)
+          ? 0xffffffff
+          : 0
     }
-
     this.node = props.__textnode
       ? renderer.createTextNode(this.props._props)
       : renderer.createNode(this.props._props)
@@ -312,7 +319,8 @@ const Element = {
       }
     }
   },
-  animate(prop, value, transition) {
+  async animate(prop, value, transition) {
+    // if current value is the same as the value to animate to, instantly resolve
     if (this.node[prop] === value) return Promise.resolve()
     // check if a transition is already scheduled or running on the same prop
     if (this.scheduledTransitions[prop]) {
@@ -345,39 +353,37 @@ const Element = {
       delay: typeof transition === 'object' ? ('delay' in transition ? transition.delay : 0) : 0,
     })
 
-    // schedule transition
-    return new Promise((resolve) => {
-      const startValue = this.node[prop]
-      this.scheduledTransitions[prop] = {
-        v: props[prop],
-        cancel: false,
-        f,
-      }
+    // capture the current value to be used in the transition start
+    const startValue = this.node[prop]
 
-      try {
-        f.start()
-          .waitUntilStarted()
-          .then((animation) => {
-            // fire transition start callback if specified
-            transition.start &&
-              typeof transition.start === 'function' &&
-              transition.start.call(this.component, this, prop, startValue)
-            // continue the chain
-            animation
-              .waitUntilStopped()
-              .then(() => delete this.scheduledTransitions[prop])
-              .then(() => {
-                // fire transition end callback if specified
-                transition.end &&
-                  typeof transition.end === 'function' &&
-                  transition.end.call(this.component, this, prop, this.node[prop])
-              })
-              .then(resolve)
-          })
-      } catch (e) {
-        Log.error(e)
-      }
-    })
+    // schedule the transition for this prop, so it can be canceled /
+    // removed if another transition for the same prop starts in the mean time
+    this.scheduledTransitions[prop] = {
+      v: props[prop],
+      cancel: false,
+      f,
+    }
+
+    // wait until the animation really starts (depending on specified delay)
+    const animation = await f.start().waitUntilStarted()
+
+    // fire transition start callback if specified
+    transition.start &&
+      typeof transition.start === 'function' &&
+      transition.start.call(this.component, this, prop, startValue)
+
+    // wait until the animation ends
+    const finished = await animation.waitUntilStopped()
+
+    // removed the prop from scheduled transitions
+    delete this.scheduledTransitions[prop]
+
+    // fire transition end callback if specified
+    transition.end &&
+      typeof transition.end === 'function' &&
+      transition.end.call(this.component, this, prop, this.node[prop])
+
+    return Promise.resolve()
   },
   destroy() {
     Log.debug('Deleting  Node', this.nodeId)